Update project to latest version of Trainworks-Reloaded.
Create a separate json folder for storing the json configuration for the mod.

Update the VS project to grab all json files under the json folder (otherwise a common mistake would be the build folder not containing all of the json files in the project)

Use AddMergedJsonFile, and add HarmonyPatch commented out (just in case)

And a small readme nudging a directory structure.